Create Webcolors.py
This program prompts the user to enter RGB values separated by commas. It then converts the input string into a tuple of integers and passes it to the get_color_name function. This function uses the rgb_to_name function from the webcolors library to get the name of the color. If the RGB values do not correspond to a known color name, it returns “Unknown”. Finally, the program prints the name of the color.