Create HSVColorSpace.py
This code loads an image and converts it from BGR to HSV color space using the (cvtColor) function from OpenCV. It then defines a range of color to detect and creates a mask using this range and the (inRange) function. The mask is applied to the original image using the (bitwise_and) function to keep only the pixels that fall within the defined color range. Finally, the result is displayed using (imshow) and (waitKey) functions.